Create the UDP socket and bind or connect to the given address:port
I had planned to always try to bind and then, depending on whether errno
was EADDRNOTAVAIL or something else, try and connect instead. But that's
a little too clever. It'll correctly report an error if another instance
is bound to the same port, but address typos will be harder to diagnose.

Better to be explicit: so the server grows an extra -l argument.