Don't bother to read the public key; it can be derived
This allows us to get rid of read_keypair and use the same function to
read the private and public key and forces minor documentation changes.

We assume that both private and public keys are 32 bytes long. That's
true now, but if crypto_box_SECRETKEYBYTES and crypto_box_PUBLICKEYBYTES
ever diverge, we'd have to do something different (pass in the array and
length to read_key, perhaps).