Clean up the code by using fewer sockaddrs
The server and client both keep their peer address in peer/peerlen. The
client can initialise this straightaway from server/srvlen; the server
has to wait for recvfrom() to initialise it. This gets rid of target
and the code to set it based on role.

This does mean that in the client, peer/peerlen are overwritten by the
recvfrom, but that's fine because we connect()ed to the server address
earlier, so the kernel will make sure we don't get packets from anyone
else. And the code becomes simpler.