Reduce the read buffer size from 64k to 2k
We expect the MTU on the TAP interfaces to be 1500 (or thereabouts). We
should therefore never get an ethernet frame larger than the MTU when
reading from the TAP device.

We'll encrypt this frame and send it as a UDP datagram. If the PMTU
between the two sides is lower than the TAP MTU, we'll get EMSGSIZE
and retry with a smaller packet size. If the ethernet frame will be
reassembled from multiple smaller write()s to the TAP device on the
other side, all will be well.

Unfortunately, that reassembly doesn't happen (verified by reading only
22 bytes at a time from the TAP fd and writing that to the UDP socket).
So our EMSGSIZE backoff code won't actually work. The only thing we can
do is to reduce the MTU of the TAP interface and hope it helps. But that
isn't something I want to do right now.