Remove all the short-write logic
We've now established that nothing good will come of writing partial
ethernet frames to the TAP device. So all of the short-write handling
logic can go away: if we can't send the whole thing, we just drop the
packet on the floor.

Experiments with ping -sNNNN show that I can send 1472-byte UDP packets
across the wire. 1473 doesn't work (but I don't get EMSGSIZE, presumably
because PMTU discovery is broken by NATs on my path). That means I can't
ignore the problem.

Reducing the TAP MTU does help, but oddly enough, reducing it to 1472
does not work. It starts working when the MTU is set to 1458, at which
point we're reading 1466-byte (ok, +8) ethernet frames from TAP. With
the MTU set to 1459, we read a 1473-byte frame, which is then dropped.
I haven't figured that out yet.

We certainly can't twiddle the interface MTU within udp_write, so we'll
have to look at some other solution. One option is to perform home-made
PMTU discovery at startup, but I don't want to do that and lengthen the
setup time. A separate program, perhaps?