Make udp_read/udp_write handle nonces
Instead of doing more buffer manipulation, we use recvmsg and sendmsg to
read and write multiple buffers. One is the nonce (or, perhaps someday,
a message header) and the other is the data. This works, save for the
nonce check.

(The other changes are all s/crypto_box_NONCEBYTES/NONCEBYTES/)