Send a keepalive packet to the server every 75 seconds
We just send a one-byte message with valid encryption. On the server,
the decryption succeeds, but we don't even need to bother to write it
to the TAP device (because the kernel would discard it anyway). This
is easier than faking a valid-but-harmless ethernet frame.

(Not properly tested, due to lack of 75s of silence.)