Now that we have the function, we can use it at startup
This eliminates the annoying warning I just added to the README about
having to send traffic from the client to the server first. (It also
eliminates a "leak" of the first counter value in the nonce, since we
always increment the nonce between tap_read and udp_write. Not that it
really mattered.)