Send keepalives from the server too
There's no especially good reason to avoid sending them, and I'm working
my way slowly towards bind()ing to INADDR_ANY on the server, so that the
client can cope with a _server_ IP address change.

If that happened right now, I expect the server would also receive
EINVAL when it tried to write to the socket. It can't avoid bind()
altogether because it needs to set the port, but INADDR_ANY should
work fine.

Is it worth doing? It certainly won't receive much testing…