Use the bundled NaCl instead of TweetNaCl
This is primarily a change to the Makefile, which uses nacl/do to build
libnacl.a and considers NaCl built if libnacl.a exists. It's not likely
that we'll build with changes to the NaCl source, so there's no support
for rebuilding NaCl.

We also have to include crypto_box.h instead of tweetnacl.h, and we can
drop our devurandom.c, because NaCl includes it.