Overhaul nonce generation and handling
This fixes a long-standing problem in tappet: if one peer was rebooted,
the tunnel would stop working until tappet was restarted on the other
peer, because the nonce counter would start again at a smaller value
(because it's derived from CLOCK_MONOTONIC, i.e., jiffies since boot).

We now use a persistent four-byte nonce prefix that is incremented at
startup, so that the MSB of the nonce is guaranteed to be larger than
any previous value, irrespective of the value of the counter.

Also, I had earlier misunderstood this part of djb's advice, reproduced
in the comment in generate_nonce():

    Choosing the nonce as a counter followed by (e.g.) 32 random bits
    helps protect some protocols against denial-of-service attacks.

If the nonce were purely a counter, an adversary could generate packets
with nonce+1, +2, etc. to keep us busy trying to decrypt bogus packets.
Appending randomness would reduce the window for such mischief, but the
four constant random bytes I was appending earlier doesn't help in this
situation. It's just useless.

The nonce is now four prefix bytes, twelve random bytes initialised at
startup, and an eight-byte nanosecond counter. The prefix and counter
are written in network byte order because the nonce is compared with
memcmp().

Many thanks again to Arnt Gulbrandsen for helping me track down and
solve the problem.