Partially revert the command line changes
This reverts changes made in the following commits:

    9967426d22fd00a6db93ee99ac888694475e434e
    fcaa68a454908049c8f5b13a14ea576ebb32a838

The idea I was pursuing was to simplify management on the listener by
using a single TAP interface and have multiple peers connect to the
same listener.

To make this work, the listener would have to identify which peer an
incoming packet is from (not too difficult, since we already track the
peer IP address) and identify which peer any given outgoing packet is
addressed *to*, which is considerably more difficult because tappet is
already used in configurations where a single tunnel is used to route
traffic for multiple addresses.

It's much easier and better to use multiple TAP interfaces than to try
to implement a half-baked routing table inside tappet.

That being the case, the command-line rearrangements become a gratuitous
change. We retain some comment changes and the argv[n] handling, but put
the arguments back in the order they used to be.